disclosure_collector_service/requirements.txt
amqp==5.3.1
appdirs==1.4.4
arelle-release==2.36.4
async-timeout==5.0.1
beautifulsoup4==4.12.3
billiard==4.2.1
blinker==1.9.0
celery==5.4.0
certifi==2024.12.14
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dart-fss==0.4.10
et_xmlfile==2.0.0
fake-useragent==2.0.3
filelock==3.16.1
Flask==3.1.0
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.5
kombu==5.4.2
lxml==5.3.0
MarkupSafe==3.0.2
numpy==2.2.1
openpyxl==3.1.5
pandas==2.2.3
pillow==11.1.0
prompt_toolkit==3.0.48
pymodm==0.4.3
pymongo==3.13.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
six==1.17.0
soupsieve==2.6
termcolor==2.3.0
tqdm==4.67.1
tzdata==2024.2
tzlocal==5.2
urllib3==2.3.0
vine==5.1.0
watchdog==6.0.0
wcwidth==0.2.13
Werkzeug==3.1.3
xmltodict==0.14.2
yaspin==3.1.0
